# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Campus Faso - Gestion du LMD",
        "description": "Documentation générée par Swagger",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/universites": {
            "get": {
                "tags": [
                    "Universités"
                ],
                "summary": "Récupérer la liste de toutes les universités",
                "operationId": "17415d2a404d6c60951e07b01d66e0aa",
                "responses": {
                    "200": {
                        "description": "Réponse de succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "libelle": {
                                                "type": "string",
                                                "example": "Université de Ouagadougou"
                                            },
                                            "abreviation": {
                                                "type": "string",
                                                "example": "UO"
                                            },
                                            "ville": {
                                                "type": "string",
                                                "example": "Ouagadougou"
                                            },
                                            "telephone": {
                                                "type": "string",
                                                "example": "+226 25 30 70 64"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "contact@uo.bf"
                                            },
                                            "site_web": {
                                                "type": "string",
                                                "example": "https://www.uo.bf"
                                            },
                                            "adresse": {
                                                "type": "string",
                                                "example": "03 BP 7021 Ouagadougou 03"
                                            },
                                            "statut": {
                                                "type": "string",
                                                "example": "active"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/universites/{id}": {
            "get": {
                "tags": [
                    "Universités"
                ],
                "summary": "Récupérer une université par son ID",
                "operationId": "d2af1d36b8b7cbc8a62e3ac15efb9c4b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'université",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Réponse de succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "libelle": {
                                                "type": "string",
                                                "example": "Université de Ouagadougou"
                                            },
                                            "abreviation": {
                                                "type": "string",
                                                "example": "UO"
                                            },
                                            "ville": {
                                                "type": "string",
                                                "example": "Ouagadougou"
                                            },
                                            "telephone": {
                                                "type": "string",
                                                "example": "+226 25 30 70 64"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "contact@uo.bf"
                                            },
                                            "site_web": {
                                                "type": "string",
                                                "example": "https://www.uo.bf"
                                            },
                                            "adresse": {
                                                "type": "string",
                                                "example": "03 BP 7021 Ouagadougou 03"
                                            },
                                            "statut": {
                                                "type": "string",
                                                "example": "active"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Université non trouvée"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/universites/{universite}/ufrs": {
            "get": {
                "tags": [
                    "Universités"
                ],
                "summary": "Récupérer la liste des UFRs appartenant à une université",
                "operationId": "a878da066ccdbc2cdeb29520f07d97ed",
                "parameters": [
                    {
                        "name": "universite",
                        "in": "path",
                        "description": "ID de l'université",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Réponse de succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "libelle": {
                                                "type": "string",
                                                "example": "UFR Sciences et Techniques"
                                            },
                                            "abreviation": {
                                                "type": "string",
                                                "example": "UFR/ST"
                                            },
                                            "statut": {
                                                "type": "string",
                                                "example": "active"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Université non trouvée"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/universites/{universite}/departements": {
            "get": {
                "tags": [
                    "Universités"
                ],
                "summary": "Récupérer la liste des départements appartenant à une université",
                "operationId": "c6685ebed1d20362dd658db83e7a11b4",
                "parameters": [
                    {
                        "name": "universite",
                        "in": "path",
                        "description": "ID de l'université",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Réponse de succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "libelle": {
                                                "type": "string",
                                                "example": "Département de Mathématiques"
                                            },
                                            "abreviation": {
                                                "type": "string",
                                                "example": "MATH"
                                            },
                                            "statut": {
                                                "type": "string",
                                                "example": "active"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Université non trouvée"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/universites/{universite}/filieres": {
            "get": {
                "tags": [
                    "Universités"
                ],
                "summary": "Récupérer la liste des filières appartenant à une université",
                "operationId": "cea6335abf9a87b538ce50dec5e22bfa",
                "parameters": [
                    {
                        "name": "universite",
                        "in": "path",
                        "description": "ID de l'université",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Réponse de succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "libelle": {
                                                "type": "string",
                                                "example": "Informatique"
                                            },
                                            "abreviation": {
                                                "type": "string",
                                                "example": "INFO"
                                            },
                                            "statut": {
                                                "type": "string",
                                                "example": "active"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Université non trouvée"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/api/universites/{universite}/niveaux": {
            "get": {
                "tags": [
                    "Universités"
                ],
                "summary": "Récupérer la liste des niveaux appartenant à une université",
                "operationId": "1209c8dee755528c77a616b3a4edcc8d",
                "parameters": [
                    {
                        "name": "universite",
                        "in": "path",
                        "description": "ID de l'université",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Réponse de succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "libelle": {
                                                "type": "string",
                                                "example": "Licence 1"
                                            },
                                            "abreviation": {
                                                "type": "string",
                                                "example": "L1"
                                            },
                                            "accessible": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "statut": {
                                                "type": "string",
                                                "example": "active"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Université non trouvée"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
...
    },
...
}